small project reorganization
