+++ lib/assertions/browser_assertions.js
@@ -3,8 +3,8 @@
 
 const assert = require('assert');
 const BrowserNotAssertions = require('./browser_not_assertions');
-const elementsAssertionUtils = require('./assertions_utils/assert_elements');
-const utils = require('./utils');
+const elementsAssertionUtils = require('./utils/assert_elements');
+const utils = require('../utils');
 
 module.exports = class BrowserAssertions {
