+++ lib/assertions/browser_not_assertions.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const assert = require('assert');
-const utils = require('./utils');
+const utils = require('../utils');
 
 function invertify(cb, msg) {
     return cb().then(() => {
