+++ lib/assertions/utils/assert_elements.js
@@ -1,5 +1,5 @@
 "use strict";
-const utils = require('../utils');
+const utils = require('../../utils');
 const assert = require('assert');
 
 const countCases = {
