+++ lib/browser/browser.js
@@ -2,7 +2,7 @@
 "use strict";
 
 const path = require('path');
-const BrowserAssertions = require('./browser_assertions');
+const BrowserAssertions = require('../assertions/browser_assertions');
 const BrowserBase = require('./browser_base');
 
 module.exports = class Browser extends BrowserBase {
@@ -19,7 +19,7 @@ module.exports = class Browser extends BrowserBase {
         return this.page.goto(url).then(() => {
             return this.page.content().then((content) => {
                 this._originalHtml = content;
-                return this.page.addScriptTag({path: path.join(__dirname, "..", "injection_scripts/wendigo_utils.js")});
+                return this.page.addScriptTag({path: path.join(__dirname, "../..", "injection_scripts/wendigo_utils.js")});
             });
         });
     }
