+++ lib/browser/browser_base.js
