+++ lib/wendigo.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const process = require('process');
-const Browser = require('./browser');
+const Browser = require('./browser/browser');
 const puppeteer = require('puppeteer');
